requirements.txt
torch==2.1.2
transformers==4.36.2
peft==0.7.1
datasets==2.14.6
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1
pyyaml==6.0.1
tensorboard==2.15.1
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
